debug energy function
