+++ source/src/rna-folding/utility/RNAQUBO.py
@@ -14,6 +14,7 @@ import pickle  # nosec
 import os
 
 from .RNAParser import RNAData
+from .RNAGeoCalc import *
 
 log = logging.getLogger()
 log.setLevel('INFO')
@@ -62,7 +63,7 @@ class RNAQUBO():
             for pkp_penalty in model_param["PKP"]:
                 for overlap_penalty in model_param["O"]:
                     for short_penalty in model_param["S"]:
-                        model_name = f"{rna_name}_{pkp_penalty}_{overlap_penalty}_{short_penalty}"
+                        model_name = f"{rna_name}+{pkp_penalty}+{overlap_penalty}+{short_penalty}+"
 
                         # check availability
                         if model_name in self.models[rna_name]['model_qubo']["qc"].keys():
@@ -75,8 +76,8 @@ class RNAQUBO():
 
 
                         start = time.time()
-                        stems_p = self._potential_stems(self.rna_data[rna_name]['rna_strand'])
-                        pks_p = self._potential_pseudoknots(stems_p[0], pkp_penalty)
+                        stems_p = self.rna_data[rna_name]['potential_stems']
+                        pks_p = potential_pseudoknots(stems_p[0], pkp_penalty)
                         ols_p = self._potential_overlaps(stems_p[0])
                         qubo_data = self._model(stems_p[0], pks_p, ols_p, stems_p[1])
                         qubo_raw = dimod.BinaryQuadraticModel(qubo_data[0], qubo_data[1], vartype = 'BINARY')
@@ -144,77 +145,6 @@ class RNAQUBO():
         with open(filename, "rb") as f:
             return pickle.load(f)  # nosec
 
-    # function to read in .fasta file and generate list of potential stems at least 3 base-pairs long:
-    def _potential_stems(self, rna_strand):
-        
-        rna = rna_strand
-        
-        matrix = np.zeros((len(rna),len(rna)))
-
-        for diag in range(0, len(matrix)):
-            for row in range(0, len(matrix)-diag):
-                col = row + diag
-                base1 = rna[row]
-                base2 = rna[col]
-                if row != col:
-                    if ((base1 == ("A" or "a")) and (base2 == ("U" or "u"))) or ((base1 == ("U" or "u")) and (base2 == ("A" or "a"))) or ((base1 == ("G" or "g")) and (base2 == ("U" or "u"))) or ((base1 == ("U" or "u")) and (base2 == ("G" or "g"))):
-                        matrix[row][col] = 2
-                    if ((base1 == ("G" or "g")) and (base2 == ("C" or "c"))) or ((base1 == ("C" or "c")) and (base2 == ("G" or "g"))):
-                        matrix[row][col] = 3
-        stems_potential = []
-        mu = 0
-
-        for row in range(0, len(matrix)):
-            for col in range (row, len(matrix)):
-                if row != col:
-                    if matrix[row][col] != 0:
-                        temp_row = row
-                        temp_col = col
-                        stem = [row+1,col+1,0,0]
-                        length_N = 0
-                        length_H = 0
-                        while (matrix[temp_row][temp_col] != 0) and (temp_row != temp_col):
-                            length_N+=1
-                            length_H+=matrix[temp_row][temp_col]
-                            temp_row+=1
-                            temp_col-=1
-                            if length_N >= 3:
-                                stem[2] = int(length_H)
-                                stem[3] = int(length_N)
-                                stems_potential.append(stem.copy())
-                        if length_H > mu:
-                            mu = length_H
-        
-        return [stems_potential, mu, rna, len(rna)]
-
-    # function to generate list of potential stem pairs that form pseudoknots:
-
-    def _potential_pseudoknots(self, stems_potential, pkp):
-
-        pseudoknots_potential = []
-        pseudoknot_penalty = pkp
-
-        for i in range(len(stems_potential)):
-            for j in range(i + 1, len(stems_potential)):
-                
-                stem1 = stems_potential[i]
-                stem2 = stems_potential[j]
-        
-                i_a = stem1[0]
-                j_a = stem1[1]
-                i_b = stem2[0]
-                j_b = stem2[1]
-        
-                pseudoknot = [i,j,1]
-        
-                if (i_a < i_b and i_b < j_a and j_a < j_b):
-            
-                    pseudoknot[2] = pseudoknot_penalty
-        
-                pseudoknots_potential.append(pseudoknot)
-                
-        return pseudoknots_potential
-
     # function to generate list of stem pairs that overlap:
 
     def _potential_overlaps(self, stems_potential):
@@ -230,11 +160,17 @@ class RNAQUBO():
         
                 overlap = [i, j, 0]
         
-                stem1_cspan1 = set(range(stem1[1]-int(stem1[3])+1, stem1[1]+1))
-                stem2_cspan1 = set(range(stem2[1]-int(stem2[3])+1, stem2[1]+1))
+                stem1_cspan1 = set(range(stem1[1]-int(stem1[2])+1, stem1[1]+1))
+                stem2_cspan1 = set(range(stem2[1]-int(stem2[2])+1, stem2[1]+1))
                 
-                stem1_cspan2 = set(range(stem1[0], stem1[0]+int(stem1[3])))
-                stem2_cspan2 = set(range(stem2[0], stem2[0]+int(stem2[3])))
+                stem1_cspan2 = set(range(stem1[0], stem1[0]+int(stem1[2])))
+                stem2_cspan2 = set(range(stem2[0], stem2[0]+int(stem2[2])))
+
+                # stem1_cspan1 = set(range(stem1[1]-int(stem1[3])+1, stem1[1]+1))
+                # stem2_cspan1 = set(range(stem2[1]-int(stem2[3])+1, stem2[1]+1))
+                
+                # stem1_cspan2 = set(range(stem1[0], stem1[0]+int(stem1[3])))
+                # stem2_cspan2 = set(range(stem2[0], stem2[0]+int(stem2[3])))
         
                 if (len(stem1_cspan1 & stem2_cspan1) != 0) or (len(stem1_cspan2 & stem2_cspan2) != 0)  or (len(stem1_cspan1 & stem2_cspan2) != 0) or (len(stem1_cspan2 & stem2_cspan1) != 0):
